pyproject.toml
[tool.poetry]
name = "todo-app"
version = "0.1.0"
description = "A simple ToDo app for your DevOps apprenticeship"
authors = []

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^2.2.5"
python-dotenv = "^0.14.0"
requests = "^2.31.0"
pytest = "^7.4.0"
gunicorn = "^21.2.0"
pymongo = "^4.6.1"
flask-dance = "^7.0.1"

[tool.poetry.dev-dependencies]

[tool.poetry.group.dev.dependencies]
mongomock = "^4.1.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
